Split recipes and recipe stats to different APIs
This refactors the logic with fetching the recipes
and recipe stats by having those as separate API
endpoints. The end-to-end execution time is now
also visualized for the PG query for listing the
recipes to better show the difference between
the PG and Redis performance.

The recipes list and the recipe detail page now
have better skeletons and empty states to prevent
the UI from flickering too much.