First set of suggested changes
Mostly changes in use of english, and removal of "Step N" from the
subtitles.

Since the change in how console works, one goes directly to a project
and then can click on Create service.

I *think* the language in the second paragraph in "Use the application"
is probably more correct - I'll need to run the application to check.